Move diffusion_discrete_set_neumann_rows to mesh/
The function diffusion_discrete_set_neumann_rows has been renamed to
'mesh_qc_set_neumann_rows' since it is called by the mesh module.
It is intended to be used for diffusion problems but intrinsically it
works directly on a qusi-cubical mesh.
This assures that there are no cyclical dependencies between modules.