Make reading of .tess files portable
The current implementation of reading a .tess file involves using
the C library function 'long ftell(FILE *)'.
On Windows, however, this causes newlines in text mode to be interpreted
as CR+LF instead of LF, thus calculating positions wrongly.
For this reason, reading of meshes is set in binary mode.