Add support for cross-compiling to Windows on Unix
This commits paves the road to porting the project to Windows.
For now it supports cross-compiling part of it on Unix-like platforms.
It has been tested by using MinGW as a cross-compiler.
The '.DEP' extension is set to '.d' for Windows as well.
(Otherwise Make may get confused.)
The option '-s' is added to 'ARFLAGS' so that an archive is created.
(Otherwise, we get an error on Windows during linking.)
A macro 'INTERPRETER' is added before any run of executables.
It is not set for native runs (thus resolved to the empty string).
However, it is set as 'INTERPRETER=wine' when Windows executables are
run on Unix-like platforms.
Installations of MinGW and Wine are shown in the 'README.md' file.
The script 'mingw.sh' is created to automate the cross-platform
(Windows on Unix) build.