Make dynamic loading cross-platform
Previously the Unix header file 'dlfcn.h' was used for dynamic loading
of plugins.
In order to port the dynamic loading logic to Windows an abstract header
file 'cmc_dynamic_library.h' has been created: it conditionally chooses
which functions to use depending on the operating system.
It is assumed that the operating system is either Windows or Unix-like.
Consequently, all functions from 'dlfcn.h' are substituted with the
corresponding functions in 'cmc_dynamic_library.h'.
As a result the text outputing diffusion demonstrations can now be
successfully run on Windows (tested on MinGW + Wine).